requirements.txt
fastapi
passlib
psycopg2
pydantic
python_jose
sqlmodel
starlette